backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
supabase==2.3.4
google-generativeai==0.3.1
python-dotenv==1.0.0
python-multipart==0.0.6
